ai_service/requirements.txt
# ====================================================================
# REQUIREMENTS.TXT — Danh sách thư viện Python cho AI Service
# ====================================================================
# Cách cài đặt: chạy lệnh `pip install -r requirements.txt` trong terminal
# ====================================================================

# --- Framework API chính ---
# FastAPI: Xây dựng các REST API tốc độ cao, hỗ trợ bất đồng bộ (async)
fastapi==0.115.0

# Uvicorn: Server ASGI chạy FastAPI trên port 8000
uvicorn==0.30.6

# --- Thư viện Khoa học Dữ liệu & Machine Learning ---
# scikit-learn: Chứa các thuật toán ML (SVM, TF-IDF, Cosine Similarity)
scikit-learn==1.5.2

# pandas: Đọc, ghi, xử lý file CSV (intent_dataset.csv, knowledge_base.csv)
pandas==2.2.3

# numpy: Thư viện tính toán số học (nền tảng cho scikit-learn)
numpy==1.26.4

# joblib: Dùng để lưu (dump) và tải (load) model AI đã train ra file .pkl
joblib==1.4.2

# --- Xử lý Ngôn ngữ Tự nhiên Tiếng Việt ---
# underthesea: Cắt từ (word segmentation) tiếng Việt chuẩn xác
# VD: "Phở Hà Nội" -> ["Phở", "Hà_Nội"] thay vì ["Phở", "Hà", "Nội"]
underthesea>=1.3.0

# --- Kết nối Redis (Phase 7 — Semantic Cache dùng chung) ---
# redis-py: Python client kết nối Redis
# Dùng để thay thế SemanticCache Python Dict (mất khi restart)
# bằng Redis-backed cache (tồn tại sau restart, chia sẻ giữa nhiều instance)
redis==5.0.0


# --- Kết nối PostgreSQL ---
# psycopg2-binary: Python client kết nối PostgreSQL
# Dùng để đọc dữ liệu món ăn và địa điểm từ database
psycopg2-binary==2.9.9
